Common seed fixture for torch tests
