+++ tests/torch/conftest.py
@@ -11,7 +11,12 @@
  limitations under the License.
 """
 import os
+import random
+
 import pytest
+
+from nncf.torch.utils import manual_seed
+
 try:
     import torch
 except: #pylint: disable=bare-except
@@ -219,6 +224,21 @@ def runs_subprocess_in_precommit():
 def cuda_ip(request):
     return request.config.getoption("--cuda-ip")
 
+
+@pytest.fixture
+def _seed():
+    if torch is not None:
+        from torch.backends import cudnn
+        cudnn.deterministic = True
+        cudnn.benchmark = False
+        torch.manual_seed(0)
+    try:
+        import numpy as np
+        np.random.seed(0)
+    except ImportError:
+        pass
+    random.seed(0)
+
 # Custom markers specifying tests to be run only if a specific option
 # is present on the pytest command line must be registered here.
 MARKS_VS_OPTIONS = {
@@ -227,3 +247,4 @@ MARKS_VS_OPTIONS = {
 
 def pytest_collection_modifyitems(config, items):
     skip_marked_cases_if_options_not_specified(config, items, MARKS_VS_OPTIONS)
+
