+++ tests/torch/quantization/test_autoq_precision_init.py
@@ -35,8 +35,6 @@ from tests.torch.quantization.test_hawq_precision_init import precision_init_dum
 from tests.torch.quantization.test_hawq_precision_init import ssd_vgg_512_test
 from tests.torch.quantization.quantization_helpers import compare_multi_gpu_dump
 from tests.torch.quantization.quantization_helpers import get_quantization_config_without_range_init
-# pylint:disable=unused-import
-from tests.torch.modules.test_rnn import _seed
 from tests.torch.test_models import inception_v3
 from tests.torch.test_models import squeezenet1_1
