+++ tests/torch/test_utils.py
@@ -10,8 +10,7 @@ from nncf.torch.initialization import DataLoaderBNAdaptationRunner
 
 from tests.torch.helpers import BasicConvTestModel, TwoConvTestModel, MockModel
 from tests.torch.quantization.test_overflow_issue_export import DepthWiseConvTestModel, EightConvTestModel
-# pylint:disable=unused-import
-from tests.torch.modules.test_rnn import _seed
+
 
 
 def compare_saved_model_state_and_current_model_state(model: nn.Module, model_state: _ModuleState):
